added 2.2.2 with @RequestParam

+++ spring_mvc/src/main/java/web/services/CarsService.java
@@ -6,5 +6,5 @@ import web.models.Car;
 import java.util.List;
 
 public interface CarsService {
-    List<Car> returnCarsByCount(int count);
+    List<Car> returnCarsByCount(Long count);
 }
