+++ spring_mvc/src/main/java/web/services/CarsServiceImpl.java
@@ -11,13 +11,13 @@ import java.util.stream.Collectors;
 public class CarsServiceImpl implements CarsService {
 
     @Override
-    public List<Car> returnCarsByCount(int count) {
+    public List<Car> returnCarsByCount(Long count) {
         List<Car> cars = new ArrayList<>();
-        cars.add(new Car(1, "Granta", "111aa"));
-        cars.add(new Car(2, "Kalina", "222cc"));
-        cars.add(new Car(3, "Priora", "333ee"));
-        cars.add(new Car(4, "Niva", "444kk"));
-        cars.add(new Car(5, "Vesta", "555oo"));
+        cars.add(new Car(1L, "Granta", "111aa"));
+        cars.add(new Car(2L, "Kalina", "222cc"));
+        cars.add(new Car(3L, "Priora", "333ee"));
+        cars.add(new Car(4L, "Niva", "444kk"));
+        cars.add(new Car(5L, "Vesta", "555oo"));
         List<Car> result = cars.stream().limit(count).collect(Collectors.toList());
         return result;
     }
